pyproject.toml
[project]
name = "bluecore-workflows"
version = "0.2.0"
description = "Blue Core Workflows using Apache Airflow"
readme = "README.md"
authors = [{ name = "Jeremy Nelson", email = "jpnelson@stanford.edu"}]
requires-python = "==3.12.*"

dependencies = [
    "pymarc",
    "apache-airflow==3.0.2", # keep version in sync with image in Dockerfile
    "apache-airflow-providers-amazon",
    "apache-airflow-providers-fab",
    "sqlalchemy",
    "typing-extensions>=4.0.0",
    "rdflib",
    "lxml",
    "folio-uuid",
    "folioclient",
    "numpy",
    "pendulum",
    "psycopg2-binary>=2.9.10",
    "boto3",
    "bluecore-models>=0.7.0",
    "xmlsec==1.3.14", # this pin can be released once a package issue with 1.3.15 is resolved
]

[dependency-groups]
dev = [
    "mypy",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-httpx",
    "types-mock",
    "requests-mock",
    "ruff>=0.9.7",
    "apache-airflow-providers-postgres",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bluecore_workflow"]

[tool.mypy]
ignore_missing_imports = true
explicit_package_bases = true

[tool.pytest.ini_options]
addopts = "-k 'not test_task_instance and not test_parse_institutional_msgs and not test_email'"


